VBliss_Blog/functions/07_GenerateFluxImage/requirements.txt
boto3
Pillow
huggingface_hub